Move devtools out of platform
