+++ src/webviews/extension-side/devTools/serviceRegistry.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 'use strict';
 
-import { IExtensionContext } from '../common/types';
-import { IServiceManager } from '../ioc/types';
+import { IExtensionContext } from '../../../platform/common/types';
+import { IServiceManager } from '../../../platform/ioc/types';
 import { addClearCacheCommand } from './clearCache';
 
 export function registerTypes(context: IExtensionContext, _serviceManager: IServiceManager, isDevMode: boolean) {
